refactor(Backend): :truck: move backend/entity to backend
