+++ backend/Cargo.toml
@@ -12,7 +12,7 @@ paste = "1.0.12"
 prost = { workspace = true }
 toml = "0.7.4"
 prost-types = { workspace = true }
-entity = { path = "./entity" }
+# entity = { path = "./entity" }
 chrono = "0.4.26"
 thiserror = "1.0.44"
 derive_builder = "0.12.0"
