--- backend/entity/.gitignore
@@ -1,1 +1,0 @@
-/pending
