--- backend/entity/src/token.rs
@@ -1,40 +1,0 @@
-//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
-
-use sea_orm::entity::prelude::*;
-
-#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "token")]
-pub struct Model {
-    #[sea_orm(primary_key)]
-    pub id: i32,
-    pub user_id: i32,
-    #[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
-    pub rand: Vec<u8>,
-    pub permission: u32,
-    #[sea_orm(column_type = "Time")]
-    pub expiry: chrono::NaiveDateTime,
-}
-
-#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
-pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::user::Entity",
-        from = "Column::UserId",
-        to = "super::user::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    User,
-}
-
-impl Related<super::user::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::User.def()
-    }
-}
-
-impl ActiveModelBehavior for ActiveModel {}
-
-impl super::DebugName for Entity {
-    const DEBUG_NAME: &'static str = "token";
-}
