+++ backend/justfile
@@ -12,7 +12,7 @@ prepare:
 entity-codegen:
     rm database/*
     sea-orm-cli migrate -u sqlite://database/backend.sqlite?mode=rwc
-    sea-orm-cli generate entity -u sqlite://database/backend.sqlite?mode=rwc -o entity/pending
+    sea-orm-cli generate entity -u sqlite://database/backend.sqlite?mode=rwc -o src/pending
 
 release-docker:
     sudo docker build --build-arg ARCH=$(uname -m) -f ./Dockerfile -t mdoj-backend ..
