+++ backend/src/endpoint/chat.rs
@@ -4,8 +4,8 @@ use super::tools::*;
 use crate::grpc::backend::chat_set_server::*;
 use crate::grpc::backend::*;
 
+use crate::entity::{chat::*, *};
 use crate::grpc::into_prost;
-use entity::{chat::*, *};
 
 impl From<i32> for ChatId {
     fn from(value: i32) -> Self {
