+++ backend/src/endpoint/mod.rs
@@ -20,9 +20,10 @@ pub mod tools {
     pub use tracing::instrument;
 }
 pub mod endpoints {
-    pub use crate::util::{filter::Filter, pager::*};
+    pub use crate::entity::DebugName;
+    pub use crate::entity::*;
+    pub use crate::util::pager::*;
     pub use crate::{fill_active_model, fill_exist_active_model, server::Server};
-    pub use entity::DebugName;
     pub use sea_orm::{
         ActiveModelTrait, ActiveValue, ColumnTrait, EntityTrait, IntoActiveModel, ModelTrait,
         PaginatorTrait, QueryFilter, QuerySelect, TransactionTrait,
