+++ backend/src/endpoint/problem.rs
@@ -3,9 +3,8 @@ use super::tools::*;
 
 use crate::grpc::backend::problem_set_server::*;
 use crate::grpc::backend::*;
-use crate::util::filter::ParentalTrait;
 
-use entity::{problem::*, *};
+use crate::entity::{problem::*, *};
 
 impl From<i32> for ProblemId {
     fn from(value: i32) -> Self {
