+++ backend/src/entity/chat.rs
@@ -1,0 +1,70 @@
+//! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
+
+use sea_orm::entity::prelude::*;
+
+use crate::endpoint::tools::Auth;
+use crate::util::error::Error;
+
+#[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
+#[sea_orm(table_name = "chat")]
+pub struct Model {
+    #[sea_orm(primary_key)]
+    pub id: i32,
+    pub user_id: i32,
+    pub problem_id: i32,
+    #[sea_orm(column_type = "Time")]
+    pub create_at: chrono::NaiveDateTime,
+    pub message: String,
+}
+
+#[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
+pub enum Relation {
+    #[sea_orm(
+        belongs_to = "super::problem::Entity",
+        from = "Column::ProblemId",
+        to = "super::problem::Column::Id",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    Problem,
+    #[sea_orm(
+        belongs_to = "super::user::Entity",
+        from = "Column::UserId",
+        to = "super::user::Column::Id",
+        on_update = "NoAction",
+        on_delete = "NoAction"
+    )]
+    User,
+}
+
+impl Related<super::problem::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::Problem.def()
+    }
+}
+
+impl Related<super::user::Entity> for Entity {
+    fn to() -> RelationDef {
+        Relation::User.def()
+    }
+}
+
+impl ActiveModelBehavior for ActiveModel {}
+
+impl super::DebugName for Entity {
+    const DEBUG_NAME: &'static str = "chat";
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, _: &Auth) -> Result<S, Error> {
+        Ok(query)
+    }
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        if let Some(perm) = auth.user_perm() {
+            if perm.can_root() || perm.can_manage_chat() {
+                return Ok(query);
+            }
+        }
+        Err(Error::Unauthenticated)
+    }
+}
