+++ backend/src/entity/contest.rs
@@ -1,6 +1,13 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
-use sea_orm::{entity::prelude::*, FromQueryResult};
+use sea_orm::QuerySelect;
+use sea_orm::{entity::prelude::*, sea_query::Alias, FromQueryResult, JoinType};
+
+use super::{user, user_contest};
+use crate::endpoint::tools::{Auth, DB};
+use crate::util::error::Error;
+
+use super::ParentalTrait;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "contest")]
@@ -89,4 +96,50 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl super::DebugName for Entity {
     const DEBUG_NAME: &'static str = "contest";
+}
+
+#[tonic::async_trait]
+impl ParentalTrait for Entity {
+    const COL_ID: Column = Column::Id;
+
+    async fn related_filter(auth: &Auth) -> Result<Select<Entity>, Error> {
+        let db = DB.get().unwrap();
+        Ok(match auth.get_user(db).await {
+            Ok(user) => user
+                .find_related(Entity)
+                .join_as(
+                    JoinType::FullOuterJoin,
+                    Relation::Hoster.def().rev(),
+                    Alias::new("own_contest"),
+                )
+                .join_as(
+                    JoinType::FullOuterJoin,
+                    user::Relation::PublicContest.def(),
+                    Alias::new("user_contest_unused"),
+                ),
+            Err(_) => Entity::find().filter(Column::Public.eq(true)),
+        })
+    }
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        if let Ok((user_id, perm)) = auth.ok_or_default() {
+            if perm.can_root() {
+                return Ok(query);
+            }
+            return Ok(query.filter(Column::Public.eq(true).or(Column::Hoster.eq(user_id))));
+        }
+        Ok(query.filter(Column::Public.eq(true)))
+    }
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        let (user_id, perm) = auth.ok_or_default()?;
+        if perm.can_root() {
+            return Ok(query);
+        }
+        if perm.can_manage_contest() {
+            return Ok(query.filter(Column::Hoster.eq(user_id)));
+        }
+        Err(Error::PermissionDeny("Can't write contest"))
+    }
 }
