+++ backend/src/entity/problem.rs
@@ -1,6 +1,11 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
-use sea_orm::{entity::prelude::*, FromQueryResult};
+use sea_orm::QuerySelect;
+use sea_orm::{entity::prelude::*, sea_query::Alias, FromQueryResult, JoinType};
+
+use super::{contest, user};
+use crate::endpoint::tools::{Auth, DB};
+use crate::util::error::Error;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel)]
 #[sea_orm(table_name = "problem")]
@@ -107,4 +112,50 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl super::DebugName for Entity {
     const DEBUG_NAME: &'static str = "problem";
+}
+
+#[tonic::async_trait]
+impl super::ParentalTrait for Entity {
+    const COL_ID: Column = Column::Id;
+
+    async fn related_filter(auth: &Auth) -> Result<Select<Entity>, Error> {
+        let db = DB.get().unwrap();
+        Ok(match auth.get_user(db).await {
+            Ok(user) => user
+                .find_linked(user::UserToProblem)
+                .join_as(
+                    JoinType::FullOuterJoin,
+                    contest::Relation::Hoster.def().rev(),
+                    Alias::new("own_problem"),
+                )
+                .join_as(
+                    JoinType::FullOuterJoin,
+                    user::Relation::PublicProblem.def(),
+                    Alias::new("problem_unused"),
+                ),
+            Err(_) => Entity::find().filter(Column::Public.eq(true)),
+        })
+    }
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        if let Ok((user_id, perm)) = auth.ok_or_default() {
+            if perm.can_root() {
+                return Ok(query);
+            }
+            return Ok(query.filter(Column::Public.eq(true).or(Column::UserId.eq(user_id))));
+        }
+        Ok(query.filter(Column::Public.eq(true)))
+    }
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        let (user_id, perm) = auth.ok_or_default()?;
+        if perm.can_root() {
+            return Ok(query);
+        }
+        if perm.can_manage_problem() {
+            return Ok(query.filter(Column::UserId.eq(user_id)));
+        }
+        Err(Error::PermissionDeny("Can't write problem"))
+    }
 }
