+++ backend/src/entity/submit.rs
@@ -1,5 +1,7 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
+use crate::endpoint::tools::Auth;
+use crate::util::error::Error;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
@@ -64,4 +66,19 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl super::DebugName for Entity {
     const DEBUG_NAME: &'static str = "submit";
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, _: &Auth) -> Result<S, Error> {
+        Ok(query)
+    }
+
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        if let Some(perm) = auth.user_perm() {
+            if perm.can_manage_submit() || perm.can_root() {
+                return Ok(query);
+            }
+        }
+        Err(Error::Unauthenticated)
+    }
 }
