+++ backend/src/entity/test.rs
@@ -1,5 +1,7 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
+use crate::endpoint::tools::Auth;
+use crate::util::error::Error;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
@@ -53,4 +55,26 @@ impl ActiveModelBehavior for ActiveModel {}
 
 impl super::DebugName for Entity {
     const DEBUG_NAME: &'static str = "testcase";
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        if let Ok((user_id, perm)) = auth.ok_or_default() {
+            if perm.can_root() {
+                return Ok(query);
+            }
+            return Ok(query.filter(Column::UserId.eq(user_id)));
+        }
+        Err(Error::PermissionDeny("Can't read testcase"))
+    }
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        let (user_id, perm) = auth.ok_or_default()?;
+        if perm.can_root() {
+            return Ok(query);
+        }
+        if perm.can_manage_problem() {
+            return Ok(query.filter(Column::UserId.eq(user_id)));
+        }
+        Err(Error::PermissionDeny("Can't write testcase"))
+    }
 }
