+++ backend/src/entity/token.rs
@@ -1,29 +1,24 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
+use crate::endpoint::tools::Auth;
+use crate::util::error::Error;
 use sea_orm::entity::prelude::*;
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
-#[sea_orm(table_name = "chat")]
+#[sea_orm(table_name = "token")]
 pub struct Model {
     #[sea_orm(primary_key)]
     pub id: i32,
     pub user_id: i32,
-    pub problem_id: i32,
+    #[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
+    pub rand: Vec<u8>,
+    pub permission: u32,
     #[sea_orm(column_type = "Time")]
-    pub create_at: chrono::NaiveDateTime,
-    pub message: String,
+    pub expiry: chrono::NaiveDateTime,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter, DeriveRelation)]
 pub enum Relation {
-    #[sea_orm(
-        belongs_to = "super::problem::Entity",
-        from = "Column::ProblemId",
-        to = "super::problem::Column::Id",
-        on_update = "NoAction",
-        on_delete = "NoAction"
-    )]
-    Problem,
     #[sea_orm(
         belongs_to = "super::user::Entity",
         from = "Column::UserId",
@@ -32,12 +27,6 @@ pub enum Relation {
         on_delete = "NoAction"
     )]
     User,
-}
-
-impl Related<super::problem::Entity> for Entity {
-    fn to() -> RelationDef {
-        Relation::Problem.def()
-    }
 }
 
 impl Related<super::user::Entity> for Entity {
@@ -49,5 +38,5 @@ impl Related<super::user::Entity> for Entity {
 impl ActiveModelBehavior for ActiveModel {}
 
 impl super::DebugName for Entity {
-    const DEBUG_NAME: &'static str = "chat";
+    const DEBUG_NAME: &'static str = "token";
 }
