+++ backend/src/entity/user.rs
@@ -1,8 +1,9 @@
 //! `SeaORM` Entity. Generated by sea-orm-codegen 0.12.6
 
+use super::{contest, problem, user_contest};
+use crate::endpoint::tools::Auth;
+use crate::util::error::Error;
 use sea_orm::entity::prelude::*;
-
-use crate::{contest, problem, user_contest};
 
 #[derive(Clone, Debug, PartialEq, DeriveEntityModel, Eq)]
 #[sea_orm(table_name = "user")]
@@ -126,4 +127,18 @@ impl Linked for UserToProblem {
 
 impl super::DebugName for Entity {
     const DEBUG_NAME: &'static str = "user";
+}
+
+impl super::Filter for Entity {
+    fn read_filter<S: QueryFilter + Send>(query: S, _: &Auth) -> Result<S, Error> {
+        Ok(query)
+    }
+
+    fn write_filter<S: QueryFilter + Send>(query: S, auth: &Auth) -> Result<S, Error> {
+        let (user_id, perm) = auth.ok_or_default()?;
+        if perm.can_root() || perm.can_manage_user() {
+            return Ok(query);
+        }
+        Ok(query.filter(Column::Id.eq(user_id)))
+    }
 }
