+++ backend/src/entity/user_contest.rs
