+++ backend/src/util/mod.rs
@@ -1,6 +1,5 @@
 pub mod auth;
 pub mod controller;
 pub mod error;
-pub mod filter;
 pub mod page;
 pub mod pager;
